server: Split lib/router.js from app.js
